feat: Factory
1.将创建annotation的逻辑从viewcontroller中抽取出来.
2.使用工厂模式进行抽取
